+++ packages/ibm-products/src/components/Cascade/Cascade.stories.js
@@ -11,18 +11,21 @@ import {
   prepareStory,
 } from '../../global/js/utils/story-helper';
 import { Cascade } from '.';
-import mdx from './Cascade.mdx';
+// import mdx from './Cascade.mdx';
 import styles from './_storybook-styles.scss'; // import index in case more files are added later.
 import { Column } from '@carbon/react';
 
 export default {
   title: getStoryTitle(Cascade.displayName),
   component: Cascade,
+  tags: ['autodocs'],
   parameters: {
     styles,
-    docs: {
+    /*
+docs: {
       page: mdx,
     },
+*/
   },
 };