+++ packages/ibm-products/src/components/CreateTearsheet/CreateTearsheet.stories.js
@@ -14,11 +14,12 @@ import { CreateTearsheet } from './CreateTearsheet';
 import { CreateTearsheetStep } from './CreateTearsheetStep';
 import { MultiStepTearsheet } from './preview-components/MultiStepTearsheet';
 import { MultiStepWithIntro } from './preview-components/MultiStepWithIntro';
-import mdx from './CreateTearsheet.mdx';
+// import mdx from './CreateTearsheet.mdx';
 
 export default {
   title: getStoryTitle(CreateTearsheet.displayName),
   component: CreateTearsheet,
+  tags: ['autodocs'],
   subcomponents: {
     CreateTearsheetStep,
   },
@@ -32,7 +33,7 @@ export default {
   },
   parameters: {
     styles,
-    docs: { page: mdx },
+    // docs: { page: mdx },
   },
 };