+++ packages/ibm-products/src/components/Datagrid/Datagrid.stories.js
@@ -27,7 +27,7 @@ import {
 } from '.';
 
 import { SelectAllWithToggle } from './Datagrid.stories/index';
-import mdx from './Datagrid.mdx';
+// import mdx from './Datagrid.mdx';
 
 import styles from './_storybook-styles.scss';
 import { DatagridActions } from './utils/DatagridActions';
@@ -38,11 +38,14 @@ import { pkg } from '../../settings';
 export default {
   title: getStoryTitle(Datagrid.displayName),
   component: Datagrid,
+  tags: ['autodocs'],
   parameters: {
     styles,
-    docs: {
+    /*
+docs: {
       page: mdx,
     },
+*/
   },
   argTypes: {
     featureFlags: {