+++ packages/ibm-products/src/components/Datagrid/utils/DatagridActions.js
@@ -14,6 +14,9 @@ import {
 } from '@carbon/react/icons';
 import {
   Button,
+  IconButton,
+  OverflowMenu,
+  OverflowMenuItem,
   ComposedModal,
   Dropdown,
   ModalBody,
@@ -68,17 +71,16 @@ export const DatagridActions = (datagridState) => {
 
   const renderFilterPanelButton = () =>
     filterProps?.variation === 'panel' && (
-      <Button
+      <IconButton
         kind="ghost"
-        hasIconOnly
-        tooltipPosition="bottom"
-        renderIcon={(props) => <Filter size={16} {...props} />}
-        iconDescription={filterProps.panelIconDescription}
+        align="bottom"
+        label={filterProps.panelIconDescription}
         className={`${blockClass}-filter-panel-open-button`}
         onClick={() => setPanelOpen((open) => !open)}
         disabled={data.length === 0}
-        tooltipAlignment="start"
-      />
+      >
+        <Filter />
+      </IconButton>
     );
 
   const [modalOpen, setModalOpen] = useState(false);