+++ packages/ibm-products/src/components/ExampleComponent/ExampleComponent.stories.js
@@ -14,7 +14,7 @@ import {
 } from '../../global/js/utils/story-helper';
 
 import { ExampleComponent } from '.';
-import mdx from './ExampleComponent.mdx';
+// import mdx from './ExampleComponent.mdx';
 import { pkg } from '../../settings';
 
 // import styles from './_storybook-styles.scss';
@@ -22,14 +22,17 @@ import { pkg } from '../../settings';
 export default {
   title: getStoryTitle(ExampleComponent.displayName),
   component: ExampleComponent,
+  tags: ['autodocs'],
   argTypes: {
     borderColor: { control: 'color' },
   },
   parameters: {
     // styles,
-    docs: {
+    /*
+docs: {
       page: mdx,
     },
+*/
   },
 };