+++ packages/ibm-products/src/components/StatusIcon/StatusIcon.stories.js
@@ -8,7 +8,7 @@
 import React from 'react';
 
 import { StatusIcon } from '.';
-import mdx from './StatusIcon.mdx';
+// import mdx from './StatusIcon.mdx';
 // import styles from './_storybook-styles.scss'; // import storybook which includes component and additional storybook styles
 import {
   getStoryTitle,
@@ -18,6 +18,7 @@ import {
 export default {
   title: getStoryTitle(StatusIcon.displayName),
   component: StatusIcon,
+  tags: ['autodocs'],
   argTypes: {
     kind: {
       control: {
@@ -52,9 +53,11 @@ export default {
   },
   parameters: {
     // styles,
-    docs: {
+    /*
+docs: {
       page: mdx,
     },
+*/
   },
 };